requirements.txt
cffi==2.0.0
cryptography==46.0.4
pycparser==3.0
PyNaCl==1.6.2
